设置模块

+++ framework/src/main/java/com/cardshop/framework/Utils/ScreenUtils.java
@@ -3,7 +3,7 @@ package com.cardshop.framework.Utils;
 import android.content.Context;
 
 public class ScreenUtils {
-    public static int calculateDpToPx(int padding_in_dp, Context context) {
+    public static int calculateDpToPx(double padding_in_dp, Context context) {
         final float scale = context.getResources().getDisplayMetrics().density;
         return (int) (padding_in_dp * scale + 0.5f);
     }
